chore: update imports and reorganize tests
