+++ package.json
@@ -12,6 +12,8 @@
     "lint-fix": "eslint --fix \"src/**/*.ts\""
   },
   "dependencies": {
+    "@apollo/server": "4.3.0",
+    "@apollo/server-plugin-landing-page-graphql-playground": "4.0.0",
     "@apollo/subgraph": "2.0.5",
     "@aws-sdk/client-eventbridge": "3.92.0",
     "@aws-sdk/client-kinesis": "3.53.0",
@@ -27,20 +29,16 @@
     "@opentelemetry/sdk-node": "0.34.0",
     "@opentelemetry/sdk-trace-node": "1.8.0",
     "@pocket-tools/apollo-cursor-pagination": "1.0.3",
-    "@pocket-tools/apollo-utils": "2.3.2",
+    "@pocket-tools/apollo-utils": "3.0.0",
     "@pocket-tools/backend-benchmarking": "1.1.0",
     "@sentry/node": "7.8.1",
     "@sentry/tracing": "7.8.1",
     "@snowplow/node-tracker": "3.3.1",
-    "apollo-server": "3.10.2",
-    "apollo-server-core": "3.10.2",
-    "apollo-server-errors": "3.3.1",
-    "apollo-server-express": "3.10.2",
-    "apollo-server-plugin-base": "3.6.2",
-    "apollo-server-plugin-response-cache": "3.7.0",
     "aws-sdk": "2.1102.0",
     "dataloader": "2.0.0",
+    "express": "4.18.2",
     "express-validator": "6.14.2",
+    "graphql": "16.6.0",
     "graphql-depth-limit": "1.1.0",
     "knex": "0.95.15",
     "locutus": "2.0.16",
