+++ src/test/graphql/mutations/savedItemsMutationService-updates.integration.ts
@@ -1,5 +1,5 @@
 import { writeClient } from '../../../database/client';
-import { gql } from 'apollo-server-express';
+import gql from 'graphql-tag';
 import chai, { expect } from 'chai';
 import chaiDateTime from 'chai-datetime';
 import sinon from 'sinon';
