+++ src/test/graphql/mutations/tagsMutationService-delete.integration.ts
@@ -1,5 +1,5 @@
 import { writeClient } from '../../../database/client';
-import { gql } from 'apollo-server-express';
+import gql from 'graphql-tag';
 import chai, { expect } from 'chai';
 import deepEqualInAnyOrder from 'deep-equal-in-any-order';
 import { ItemsEventEmitter } from '../../../businessEvents';
