+++ src/test/graphql/queries/suggested-tags.integration.ts
@@ -1,5 +1,5 @@
 import { readClient } from '../../../database/client';
-import { gql } from 'apollo-server-express';
+import gql from 'graphql-tag';
 import chai, { expect } from 'chai';
 import chaiDateTime from 'chai-datetime';
 import { getServer } from '../testServerUtil';
